feat(api/clients/v2): Support V3 cert type for compatibility with new Router && CertVerifier interfaces - address PR feedback & remove legacy verifier / v2 cert type

+++ api/clients/v2/payloadretrieval/validator_payload_retriever.go
@@ -72,7 +72,7 @@ func (pr *ValidatorPayloadRetriever) GetPayload(
 		blob, err := pr.retrieveBlobWithTimeout(
 			ctx,
 			*blobKey,
-			uint16(eigenDACert.BlobVersion()),
+			eigenDACert.BlobVersion(),
 			*commitment,
 			eigenDACert.ReferenceBlockNumber(),
 			quorumID)
